Modificaciones de prueba, solo repaso

+++ 04_Sentencia_condicional_(if_y_switch)/Ejercicios/T04Ejercicio17.java
@@ -10,13 +10,13 @@ import java.util.Scanner;
 public class T04Ejercicio17 {
   public static void main(String[] args) {
     
-    int numero;
+    double numero;
 
     Scanner teclado = new Scanner(System.in);
 
     System.out.println("");
     System.out.print("Diga un número entero: ");
-    numero = teclado.nextInt();
+    numero = teclado.nextDouble();
     System.out.println("");
     
     teclado.close();
